requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Django==5.1.3
django-allauth==65.3.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
sentry-sdk==2.19.0
sqlparse==0.5.2
urllib3==2.2.3
whitenoise==6.8.2
